requirements.txt
langchain_community
langchain_core
langchain_Groq
pypdf
validators ###which helps you to validate the url
youtube_transcript_api ###to fetch the transcript
streamlit
unstructered
pytube